refactor(renderer) split out HTML5 templates
This breaks the former html5 renderer into a generic backend
SGML emitter, and an HTML5 specific flavor of that.  This
should facilitate other SGML-derived backends such as docbook,
XHTML, and semantic HTML.  It also lays the foundation for
other output formats like EPUB, markup-free ASCII, and perhaps
troff style output.

As part of this refactoring work, a bug related to the Context
being passed by value instead of by reference is fixed.

Also, some other minor style fixups, and improvements were made.
For example, bytes.NewBuffer(nil) was changed to &bytes.Buffer{}
(which potentially is also faster) and the use of type aliases
to improve readability.

Tests were unmodified except fixes needed to adjust for changes
to the package paths.

To build derived implementations, the html5 packages exports
a Templates() function that yields up a copy of the string
templates that it is using.