refactor(renderer): Use inline UTF-8 (optionally)
This adds an option to render UTF-8 as is, doing the final
conversion to HTML entities only at the final step.  This
is more likely to result in correct encoding as they remain
unadulterated until the final rendering step.

Furthermore, some renderers may wish to allow Unicode to
be encoded as is.

While here... fixed the encoding of copyright.